refactor: 30분 전에 eta 활성화되는 비즈니스 로직을 도메인으로 이동
