+++ android/app/src/main/java/com/mulberry/ody/data/repository/DefaultMatesEtaRepository.kt
@@ -4,14 +4,14 @@ import android.content.Context
 import com.mulberry.ody.data.local.db.MateEtaInfoDao
 import com.mulberry.ody.data.local.db.OdyDataStore
 import com.mulberry.ody.data.local.entity.eta.MateEtaInfoEntity
-import com.mulberry.ody.data.local.service.EtaDashboard
 import com.mulberry.ody.data.local.service.EtaDashboardService
+import com.mulberry.ody.domain.common.toMilliSeconds
 import com.mulberry.ody.domain.model.MateEtaInfo
+import com.mulberry.ody.domain.model.MeetingDateTime
 import com.mulberry.ody.domain.repository.ody.MatesEtaRepository
 import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
-import java.time.LocalDateTime
 import javax.inject.Inject
 
 class DefaultMatesEtaRepository
@@ -19,7 +19,6 @@ class DefaultMatesEtaRepository
     constructor(
         @ApplicationContext private val context: Context,
         private val matesEtaInfoDao: MateEtaInfoDao,
-        private val etaDashboard: EtaDashboard,
         private val odyDataStore: OdyDataStore,
     ) : MatesEtaRepository {
         override fun fetchMatesEtaInfo(meetingId: Long): Flow<MateEtaInfo?> {
@@ -32,9 +31,11 @@ class DefaultMatesEtaRepository
 
         override fun openEtaDashboard(
             meetingId: Long,
-            meetingDateTime: LocalDateTime,
+            meetingDateTime: MeetingDateTime,
         ) {
-            etaDashboard.open(meetingId, meetingDateTime)
+            val meetingTimeMills = meetingDateTime.dateTime.toMilliSeconds()
+            val serviceIntent = EtaDashboardService.getIntent(context, meetingId, meetingTimeMills, isOpen = true)
+            context.startForegroundService(serviceIntent)
         }
 
         override suspend fun closeEtaDashboard(meetingId: Long) {
