+++ android/app/src/main/java/com/mulberry/ody/domain/repository/ody/MatesEtaRepository.kt
@@ -1,6 +1,7 @@
 package com.mulberry.ody.domain.repository.ody
 
 import com.mulberry.ody.domain.model.MateEtaInfo
+import com.mulberry.ody.domain.model.MeetingDateTime
 import kotlinx.coroutines.flow.Flow
 import java.time.LocalDateTime
 
@@ -11,7 +12,7 @@ interface MatesEtaRepository {
 
     fun openEtaDashboard(
         meetingId: Long,
-        meetingDateTime: LocalDateTime,
+        meetingDateTime: MeetingDateTime,
     )
 
     suspend fun closeEtaDashboard(meetingId: Long)
