+++ android/app/src/main/java/com/mulberry/ody/presentation/feature/room/MeetingRoomViewModel.kt
@@ -201,7 +201,7 @@ class MeetingRoomViewModel
 
         fun navigateToEtaDashboard() {
             viewModelScope.launch {
-                if (!meeting.value.isEtaAccessible()) {
+                if (!meeting.value.isEtaOpenTime()) {
                     _inaccessibleEtaEvent.emit(Unit)
                     return@launch
                 }
